Changed SignedUpAt from int32 to int 64 to reslove an issue in fleet-api

+++ user.go
@@ -25,7 +25,8 @@ type User struct {
 	Pseudonym              string                 `json:"pseudonym,omitempty"`
 	Avatar                 *UserAvatar            `json:"avatar,omitempty"`
 	LocationData           *LocationData          `json:"location_data,omitempty"`
-	SignedUpAt             int32                  `json:"signed_up_at,omitempty"`
+	SignedUpAt             int64                  `json:"signed_up_at,omitempty"`
+	// SignedUpAt             int32                  `json:"signed_up_at,omitempty"`
 	RemoteCreatedAt        int32                  `json:"remote_created_at,omitempty"`
 	LastRequestAt          int32                  `json:"last_request_at,omitempty"`
 	CreatedAt              int32                  `json:"created_at,omitempty"`
